Update Components and Pages to be inside of src
