--- front-end/Components/Header.js
@@ -1,4 +1,0 @@
-
-export default function Header() {
-  return;
-}
