--- front-end/Components/NavBar.js
@@ -1,5 +1,0 @@
-
-
-export default function NavBar() {
-    return
-}
