--- front-end/Pages/Edit.js
@@ -1,5 +1,0 @@
-
-
-export default function Edit() {
-  return;
-}
