--- front-end/Pages/FourOFour.js
@@ -1,5 +1,0 @@
-
-
-export default function FourOFour() {
-  return;
-}
