--- front-end/Pages/Home.js
@@ -1,4 +1,0 @@
-
-export default function Home() {
-  return;
-}
