--- front-end/Pages/Index.js
@@ -1,4 +1,0 @@
-
-export default function Index() {
-  return;
-}
