--- front-end/Pages/New.js
@@ -1,4 +1,0 @@
-
-export default function New() {
-  return;
-}
