--- front-end/Pages/Show.js
@@ -1,4 +1,0 @@
-
-export default function Show() {
-  return;
-}
