+++ front-end/src/Components/Header.js
@@ -1,0 +1,4 @@
+
+export default function Header() {
+  return;
+}
