+++ front-end/src/Components/NavBar.js
@@ -1,0 +1,5 @@
+
+
+export default function NavBar() {
+    return
+}
