+++ front-end/src/Pages/Edit.js
@@ -1,0 +1,5 @@
+
+
+export default function Edit() {
+  return;
+}
