+++ front-end/src/Pages/FourOFour.js
@@ -1,0 +1,5 @@
+
+
+export default function FourOFour() {
+  return;
+}
