+++ front-end/src/Pages/Home.js
@@ -1,0 +1,4 @@
+
+export default function Home() {
+  return;
+}
