+++ front-end/src/Pages/Index.js
@@ -1,0 +1,4 @@
+
+export default function Index() {
+  return;
+}
