+++ front-end/src/Pages/New.js
@@ -1,0 +1,4 @@
+
+export default function New() {
+  return;
+}
