+++ front-end/src/Pages/Show.js
@@ -1,0 +1,4 @@
+
+export default function Show() {
+  return;
+}
